% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSModel.R
\name{BlackScholesCallPrice}
\alias{BlackScholesCallPrice}
\title{Black-Scholes price of European call option}
\usage{
BlackScholesCallPrice(t = 0, T, S, K, r, q = 0, sigma)
}
\arguments{
\item{t}{Current time (default t=0)}

\item{T}{Maturity date}

\item{S}{Current asset price}

\item{K}{Strike price}

\item{r}{Risk-free interest rate}

\item{q}{Dividend yield}

\item{sigma}{Volatility}
}
\value{
Price of European call option
}
\description{
Black-Scholes price of European call option
}
