% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSModel.R
\name{BlackScholesPutPrice}
\alias{BlackScholesPutPrice}
\title{Black-Scholes price of European put option}
\usage{
BlackScholesPutPrice(t = 0, T, S, K, r, q = 0, sigma)
}
\arguments{
\item{t}{Current time (default t=0)}

\item{T}{Maturity date}

\item{S}{Current asset price}

\item{K}{Strike price}

\item{r}{Risk-free interest rate}

\item{q}{Dividend yield}

\item{sigma}{Volatility test}
}
\value{
Price of European put option
}
\description{
Black-Scholes price of European put option
}
