% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEuropeanFDM.R
\name{errorTable}
\alias{errorTable}
\title{Table of errors for different N and M}
\usage{
errorTable(
  N,
  M,
  scheme = explicitFDMEuropeanPutError,
  view = TRUE,
  theta = NULL
)
}
\arguments{
\item{N}{Sequence of spatial grid points}

\item{M}{Sequence of time steps}

\item{view}{View dataframe}

\item{schem}{FDM scheme to be applied}
}
\value{
Outputs table of error for different
N and M in latex
}
\description{
Table of errors for different N and M
}
