% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDMEuropeanPrices.R
\name{explicitFDMEuropeanPut}
\alias{explicitFDMEuropeanPut}
\title{European put price using FDM}
\usage{
explicitFDMEuropeanPut(sigma, r, T, K, Smax, N, M)
}
\arguments{
\item{sigma}{Volatility}

\item{r}{Risk-free interest rate}

\item{K}{Strike price}

\item{Smax}{Truncation of asset price}

\item{N}{Number of spatial intervals}

\item{M}{Number of time-steps}
}
\value{
Vector of prices e.g. V[1,] give the prices at time 0 for each S
}
\description{
European put price using FDM
}
