% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceEuropeanFDM.R
\name{explicitFDMEuropeanPutError}
\alias{explicitFDMEuropeanPutError}
\title{Explicit finite-difference scheme error for ATM
European put options}
\usage{
explicitFDMEuropeanPutError(N, M)
}
\arguments{
\item{N}{Number of spatial intervals}

\item{M}{Number of time-steps}
}
\value{
The error between the explicit FDM price
at time 0 for an ATM put and its Black-Scholes
price
}
\description{
Explicit finite-difference scheme error for ATM
European put options
}
