% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceEuropeanFDM.R
\name{implicitFDMEuropeanPut}
\alias{implicitFDMEuropeanPut}
\title{Implicit finite-difference scheme for pricing
European put options}
\usage{
implicitFDMEuropeanPut(sigma, r, T, K, Smax, N, M)
}
\arguments{
\item{sigma}{Volatility}

\item{r}{Risk-free interest rate}

\item{K}{Strike price}

\item{Smax}{Truncation of asset price}

\item{N}{Number of spatial intervals}

\item{M}{Number of time-steps}
}
\value{
Vector of prices V. V[m,] give the prices
at time-step m for each S
}
\description{
Implicit finite-difference scheme for pricing
European put options
}
