% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceEuropeanFDM.R
\name{implicitFDMEuropeanPutError}
\alias{implicitFDMEuropeanPutError}
\title{Implicit finite-difference scheme error for ATM
European put options}
\usage{
implicitFDMEuropeanPutError(N, M)
}
\arguments{
\item{N}{Number of spatial intervals}

\item{M}{Number of time-steps}
}
\value{
The error between the implicit FDM price
at time 0 for an ATM put and its Black-Scholes
price
}
\description{
Implicit finite-difference scheme error for ATM
European put options
}
