% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEuropeanFDM.R
\name{plotErrorMultiple}
\alias{plotErrorMultiple}
\title{Error plots for implicit FDM scheme at time 0 for
different N and M}
\usage{
plotErrorMultiple(P1, P2, P3, P4)
}
\arguments{
\item{P1}{Error plot 1 list}

\item{P2}{Error plot 3 list}

\item{P4}{Error plot 4 list}
}
\value{
A plot of the difference between the any FDM
solution and the Black-Scholes price, against the asset price at
time 0, for 4 pairs of (N,M)
}
\description{
Error plots for implicit FDM scheme at time 0 for
different N and M
}
