% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEuropeanFDM.R
\name{plotFDMError}
\alias{plotFDMError}
\title{Error plots for FDM scheme at time 0}
\usage{
plotFDMError(
  N,
  M,
  scheme = explicitFDMEuropeanPut,
  error = TRUE,
  return = FALSE,
  plotV = FALSE,
  theta = NULL
)
}
\arguments{
\item{N}{Number of spatial grid points}

\item{M}{Number of time steps}

\item{error}{If TRUE, produces error plot}
}
\value{
A plot of the difference between the a FDM solution
and the Black-Scholes price, against the asset price at
time 0, for any N and M
}
\description{
Error plots for FDM scheme at time 0
}
