% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEuropeanFDM.R
\name{plotsExplicitFDM}
\alias{plotsExplicitFDM}
\title{Error plots for FDM scheme at time 0}
\usage{
plotsExplicitFDM(N, M, error = TRUE)
}
\arguments{
\item{N}{Number of spatial grid points}

\item{M}{Number of time steps}

\item{error}{If TRUE, produces error plot}
}
\value{
A plot of the difference between the FDM solution
and the Black-Scholes price, against the asset price at
time 0, for any N and M
}
\description{
Error plots for FDM scheme at time 0
}
