% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceEuropeanFDM.R
\name{thetaFDMEuropeanPut}
\alias{thetaFDMEuropeanPut}
\title{Theta finite-difference scheme for pricing
European put options}
\usage{
thetaFDMEuropeanPut(sigma, r, T, K, Smax, N, M, theta = 0.5)
}
\arguments{
\item{sigma}{Volatility}

\item{r}{Risk-free interest rate}

\item{K}{Strike price}

\item{Smax}{Truncation of asset price}

\item{N}{Number of spatial intervals}

\item{M}{Number of time-steps}

\item{theta}{default = 0.5 (Crank-Nicolson)}
}
\value{
Vector of prices V. V[m,] give the prices
at time-step m for each S
}
\description{
Theta finite-difference scheme for pricing
European put options
}
