% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priceEuropeanFDM.R
\name{thetaFDMEuropeanPutError}
\alias{thetaFDMEuropeanPutError}
\title{Theta finite-difference scheme error for ATM
European put options}
\usage{
thetaFDMEuropeanPutError(N, M, theta)
}
\arguments{
\item{N}{Number of spatial intervals}

\item{M}{Number of time-steps}

\item{theta}{default = 0.5 (Crank-Nicolson)}
}
\value{
The error between the theta FDM price
at time 0 for an ATM put and its Black-Scholes
price
}
\description{
Theta finite-difference scheme error for ATM
European put options
}
