% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operationsMatrices.R
\name{thomasAlgorithmSolver}
\alias{thomasAlgorithmSolver}
\title{Thomas algorithm}
\usage{
thomasAlgorithmSolver(A, b)
}
\arguments{
\item{A}{Square matrix}

\item{b}{A vector}
}
\value{
The solution x of the equation
Ax = b
}
\description{
Thomas algorithm
}
