% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operationsMatrices.R
\name{tridiagonalMatrix}
\alias{tridiagonalMatrix}
\title{Making a tridiagonal matrix}
\usage{
tridiagonalMatrix(a, b, c)
}
\arguments{
\item{a}{Sub-diagonal}

\item{b}{Main diagonal}

\item{c}{Super-diagonal}
}
\value{
A tridiagonal matrix
}
\description{
Making a tridiagonal matrix
}
