% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDMEuropeanPlots.R
\name{plotFDMV}
\alias{plotFDMV}
\title{Plotting function for FDM solution}
\usage{
plotFDMV(S, V, BS, m = 1)
}
\arguments{
\item{S}{Asset price}

\item{V}{Matrix of FDM values}

\item{BS}{Black-Scholes prices}

\item{m}{Time-step}
}
\value{
NA
}
\description{
Plotting function for FDM solution
}
